Don't flush the peer list between synchronisations
localhost is added to the peer list at the start of the program.
For each file to sync, the peer list is reused. It will be flushed after
all args have been processed.

This commit also fixes a small "bug" where 'syncfile()' doesn't wait for
all its chilren to return.
This means we won't sync the next file until the current one isn't
finished synchronising. Not sure wether or not this is a good thing, but
at least the logs appear in the "natural" order now.